LPS-72125 Move LiferayEndpoint/Dispatcher & GroupScoped to a liferay specific module
To decouple vulcan-application & vulcan-api from Liferay
