+++ modules/apps/foundation/vulcan/vulcan-api/bnd.bnd
@@ -4,7 +4,6 @@ Bundle-Version: 1.0.0
 Export-Package:\
 	com.liferay.vulcan.contributor,\
 	com.liferay.vulcan.endpoint,\
-	com.liferay.vulcan.liferay.scope,\
 	com.liferay.vulcan.pagination,\
 	com.liferay.vulcan.representor,\
 	com.liferay.vulcan.representor.builder,\
