+++ modules/apps/foundation/vulcan/vulcan-liferay-api/bnd.bnd
@@ -1,0 +1,3 @@
+Bundle-Name: Liferay Vulcan Liferay API
+Bundle-SymbolicName: com.liferay.vulcan.liferay.api
+Bundle-Version: 1.0.0
