+++ modules/apps/foundation/vulcan/vulcan-liferay-api/build.gradle
@@ -1,0 +1,15 @@
+sourceCompatibility = "1.8"
+targetCompatibility = "1.8"
+
+dependencies {
+	provided group: "com.liferay", name: "com.liferay.osgi.service.tracker.collections", version: "2.0.0"
+	provided group: "com.liferay.portal", name: "com.liferay.portal.kernel", version: "2.0.0"
+	provided group: "javax.portlet", name: "portlet-api", version: "2.0"
+	provided group: "javax.servlet", name: "javax.servlet-api", version: "3.0.1"
+	provided group: "javax.ws.rs", name: "javax.ws.rs-api", version: "2.0.1"
+	provided group: "org.apache.cxf", name: "cxf-core", version: "3.1.9"
+	provided group: "org.apache.cxf", name: "cxf-rt-frontend-jaxrs", version: "3.1.9"
+	provided group: "org.osgi", name: "org.osgi.core", version: "6.0.0"
+	provided group: "org.osgi", name: "org.osgi.service.component.annotations", version: "1.3.0"
+	provided project(":apps:foundation:vulcan:vulcan-api")
+}
