+++ modules/apps/foundation/vulcan/vulcan-liferay-api/src/main/java/com/liferay/vulcan/liferay/scope/GroupScoped.java
